src/rd_upsampling/requirements.txt
absl-py
astunparse
cachetools
certifi
charset-normalizer
click
docker-pycreds
flatbuffers
gast
gin
gin-config
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib_metadata
joblib
keras
libclang
Markdown
MarkupSafe
matplotlib
numpy
oauthlib
opt_einsum
packaging
platformdirs
psutil
pyasn1
pyasn1_modules
PyYAML
requests
requests-oauthlib
rsa
scikit-learn
scipy
sentry-sdk
setproctitle
six
smmap
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
termcolor
threadpoolctl
tqdm
typing_extensions
urllib3
# In remote server, recommand version of wandb==0.15.11
wandb
Werkzeug
wrapt
zipp
